Enhancement: convert method of file processing
- Previous version used C++'s std::locale functionality to
  delimit words while reading from the file. This method is super
  memory efficient, and keeps the logic extremely simple.
- Updated to read data into a buffer and process the buffer using
  RegEx functionality to locate words.

Note: The former std:locale version may have been reading bytes
from the file in such a way that unicode data may have been
individually split up; f.e an accented letter in Unicode can
be represented as the letter character + the accent characters.
When reading at the byte level, the letter may have been read in
and considered a character while the accent was discarded. Qt
is able to understand Unicode properly so the letter and accent
are then combined together into a single character. Either method
of counting could be considered correct depending on what view
of the data one wishes to take.